Reach >90% code coverage

+++ generators/core/index.js
@@ -1,28 +1,16 @@
 const path = require('path')
-const fs = require('fs')
 const Generator = require('yeoman-generator')
 const config = require('config')
 const hostile = require('hostile')
-
-function directoryExists (path) {
-  // fs.stat requires a callback, so we have to Promise
-  return new Promise((resolve) => {
-    fs.stat(path, (err, stats) => {
-      if (err) {
-        return resolve(false)
-      }
-      if (stats.isDirectory()) {
-        resolve(true)
-      } else {
-        resolve(false)
-      }
-    })
-  })
-}
+const {
+  directoryExists,
+  fileExists
+} = require('../../utils/dir')
 
 module.exports = class extends Generator {
   initializing () {
     this.log('initializing')
+    this.sourceRoot(path.join(__dirname, 'templates'))
   }
   prompting () {
     this.log('prompting')
@@ -51,52 +39,53 @@ module.exports = class extends Generator {
   }
 
   writingRepos () {
-    this.log('writing')
+    if (this.options['skip-install'] !== true) {
+      this.log('Writing repositories')
 
-    const pathApi = path.resolve(this.answers.path, 'api')
-    const pathWeb = path.resolve(this.answers.path, 'web')
-    const gitBase = `git@github.com:${this.answers.gitUser}`
+      const pathApi = path.resolve(this.answers.path, 'api')
+      const pathWeb = path.resolve(this.answers.path, 'web')
+      const gitBase = `git@github.com:${this.answers.gitUser}`
 
-    this._cloneRepo(`${gitBase}/${config.get('repositories.api')}`, pathApi)
-    this._cloneRepo(`${gitBase}/${config.get('repositories.web')}`, pathWeb)
+      this._cloneRepo(`${gitBase}/${config.get('repositories.api')}`, pathApi)
+      this._cloneRepo(`${gitBase}/${config.get('repositories.web')}`, pathWeb)
+
+      this.log('Writing .env files')
+      fileExists(this.destinationPath('./api/app/.env.example')) && this.fs.copy(
+        this.destinationPath('./api/app/.env.example'),
+        this.destinationPath('./api/app/.env')
+      )
+      fileExists(this.destinationPath('./web/.env.example')) && this.fs.copy(
+        this.destinationPath('./web/.env.example'),
+        this.destinationPath('./web/.env')
+      )
+    }
   }
 
   writingTemplates () {
     this.log('writing templates')
     this.fs.copyTpl(
-      this.templatePath('./core/package.json'),
+      this.templatePath('package.json'),
       this.destinationPath('./package.json'),
       this.answers
     )
     this.fs.copyTpl(
-      this.templatePath('./core/docker-compose.yml'),
+      this.templatePath('docker-compose.yml'),
       this.destinationPath('./docker-compose.yml'),
       this.answers
     )
     this.fs.copy(
-      this.templatePath('./core/docker'),
+      this.templatePath('docker'),
       this.destinationPath('./docker')
-    )
-
-    this.log('Writing .env files')
-    this.fs.copy(
-      this.destinationPath('./api/app/.env.example'),
-      this.destinationPath('./api/app/.env')
-    )
-    this.fs.copy(
-      this.destinationPath('./web/.env.example'),
-      this.destinationPath('./web/.env')
     )
   }
 
   end () {
     hostile.get(false, (err, lines) => {
       if (err) {
-        throw new Error('Uh oh, looks like I had an issue reading the hosts file')
+        this.log('Uh oh, looks like I had an issue reading the hosts file')
       }
 
       const configured = lines.reduce((memo, line) => {
-        console.log(line)
         if (/sprucebot.com/.test(line[1])) {
           memo[line[1]] = true
         }
@@ -111,8 +100,8 @@ module.exports = class extends Generator {
     this.log('end')
   }
 
-  async _cloneRepo (repo, path) {
-    const exists = await directoryExists(path)
+  _cloneRepo (repo, path) {
+    const exists = directoryExists(path)
     if (exists) {
       this.log(`Uh oh, looks like you already installed something at ${path}!`)
     } else {
